requirements.txt
opencv-python
pillow
pytesseract
numpy
